+++ server/api/src/main/java/com/twb/pokerapp/dto/table/TableDTO.java
@@ -3,6 +3,7 @@ package com.twb.pokerapp.dto.table;
 import com.twb.pokerapp.domain.enumeration.GameType;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -14,6 +15,6 @@ public class TableDTO {
     private Integer totalRounds;
     private Integer minPlayers;
     private Integer maxPlayers;
-    private Double minBuyin;
-    private Double maxBuyin;
+    private BigDecimal minBuyin;
+    private BigDecimal maxBuyin;
 }
